Rename PIDController to PidController, remove all but gains logic
